[CXF-7555] Moving SpringViewResolverProvider to .springmvc. subpackage
