+++ rt/frontend/jaxrs/src/main/java/org/apache/cxf/jaxrs/springmvc/Messages.properties
